Add script to generate a heat map of Outreachy applicant locations.
The script uses the choropleth plot type from plotly to generate a map
of where Outreachy applicants are from.

For now it writes the whole chart to an HTML file, which includes the
plotly javascript. In the future we might add additional plots to the
page, and include the javascript at the end of the file.