Update posters for new initial application deadline and logos.
The Outreachy initial application deadline is now at the end of February
or the end of September. Applicants still have to make contributions in
March and October, but they won't be able to submit an initial
application during that time. Emphasize the earliest possible deadline
to ensure everyone applies during that time.

Remove the git logo (they haven't participated in several rounds) and
the Kubernetes logo (they are inconsistent about participating). Replace
it with the LibreHealth logo, since they have participated in the last
three rounds. This also gives us an interesting project to talk about
that shows the direct impact internships can have to make the world
better.