Add script to track respondents to the longitudinal survey.
Note that for some reason, libreoffice writes the CSV file with "very
long lines" but this script creates a file with default line endings. If
you're saving the temporary file to update the old participant CSV,
you'll have to open in LibreOffice, change something, save it, undo that
change, and save it again. Then the git diff will look reasonable.