Rewrite script to generate semi-personalized mass emails.
Revise the script to be more generic, taking positional arguments for
things like from headers and a subject line.

Make the script parse a message body file, and replace the phrase "Hi
NAME," with either the recipient's first name or the phrase "Greetings,"

The first name is assumed to be the first word in the Name row in the
recipient CSV file. That's not always true, depending on which culture
the recipient lives in. You may need to either modify the CSV file, or
edit the greetings line if the culture lists last name as the first
word of a full name.

Add more descriptive help text.

Update the copyright to reflect my name change and official Conservancy
email address. As critters say, 🎶 "It's been a while..." 🎶