Make back of shirt Outreachy text color match front of shirt
I played around with making the text on the back of the shirt be
transparent (meaning the black shirt shows through on the letters).
However, based on the sample I got, I'm afraid the print will start to
peel if there are too many edges on large ink print spots. So instead
we'll match the color of the text on the front of the shirt, which is
the same as the manatee outline.